feat: add Docker support and update scripts
This commit adds Docker configurations for easier environment setup and deployment. Additionally, it updates the scripts in the package.json file to start the project with PM2 in a production environment and includes a dev script with nodemon for the development environment. A logging statement has also been introduced to print the root path when the server starts, and the server is set to listen to "0.0.0.0" instead of "localhost".